refactor: update follow
