+++ src/app/mod.rs
@@ -1,4 +1,3 @@
 pub mod article;
 pub mod comment;
-pub mod follow;
 pub mod tag;
