+++ src/appv2/features/follow/mod.rs
@@ -1,0 +1,1 @@
+pub mod entities;
