+++ src/appv2/features/mod.rs
@@ -1,4 +1,5 @@
 pub mod favorite;
+pub mod follow;
 pub mod healthcheck;
 pub mod profile;
 pub mod user;
